implement data accumulation
