+++ qmcpy/stopping_criterion/cub_qmc_bayes_lattice_g.py
@@ -1,5 +1,5 @@
 from ._stopping_criterion import StoppingCriterion
-from ..accumulate_data import LDTransformBayesData, OutParams, OutOptParams
+from ..accumulate_data.ld_transform_bayes_data import LDTransformBayesData, OutParams, OutOptParams
 from ..discrete_distribution import Lattice
 from ..true_measure import Gaussian
 from ..integrand import Keister
@@ -77,12 +77,22 @@ class CubBayesLatticeG(StoppingCriterion):
     """
 
     def __init__(self, integrand, abs_tol=1e-2, rel_tol=0,
-                 n_init=2 ** 8, n_max=2 ** 22, alpha=0.01):
+                 n_init=2 ** 8, n_max=2 ** 22, alpha=0.01, ptransform = 'C1sin'):
         # Set Attributes
         self.absTol = abs_tol
         self.relTol = rel_tol
-        self.m_min = log2(n_init)
-        self.m_max = log2(n_max)
+        m_min = log2(n_init)
+        m_max = log2(n_max)
+        if m_min%1 != 0. or m_min < 8 or m_max%1 != 0:
+            warning_s = '''
+                n_init and n_max must be a powers of 2.
+                n_init must be >= 2^8.
+                Using n_init = 2^10 and n_max=2^23.'''
+            warnings.warn(warning_s, ParameterWarning)
+            m_min = 10.
+            m_max = 23.
+        self.m_min = m_min
+        self.m_max = m_max
         self.n_min = n_init  # number of samples to start with = 2^mmin
         self.n_max = n_max  # max number of samples allowed = 2^mmax
         self.alpha = alpha  # p-value, default 0.1%.
@@ -94,7 +104,7 @@ class CubBayesLatticeG(StoppingCriterion):
         self.useGradient = False  # If true uses gradient descent in parameter search
         self.oneTheta = True  # If true use common shape parameter for all dimensions
         # else allow shape parameter vary across dimensions
-        self.ptransform = 'C1sin'  # periodization transform
+        self.ptransform = ptransform  # periodization transform
         self.stopAtTol = True  # automatic mode: stop after meeting the error tolerance
         self.arbMean = True  # by default use zero mean algorithm
         self.stopCriterion = 'MLE'  # Available options {'MLE', 'GCV', 'full'}
@@ -118,8 +128,6 @@ class CubBayesLatticeG(StoppingCriterion):
         self.dscAll = []
         self.s_All = []
 
-        self.mvec = np.arange(self.m_min, self.m_max, dtype=int)
-
         # Credible interval : two-sided confidence, i.e., 1-alpha percent quantile
         if self.fullBayes:
             # degrees of freedom = 2^mmin - 1
@@ -127,13 +135,10 @@ class CubBayesLatticeG(StoppingCriterion):
         else:
             self.uncert = -gaussnorm.ppf(self.alpha / 2)
 
-        #  generator for the Lattice points
-        self.gen_vec = self.get_lattice_gen_vec(self.dim)
-        self.ff = CubBayesLatticeG.doPeriodTx(self.integrand.f, self.ptransform)  # integrand after the periodization transform
-
         # Verify Compliant Construction
         distribution = integrand.measure.distribution
-        allowed_levels = 'single'
+        self.distribution = distribution
+        allowed_levels = ['single']
         allowed_distribs = ["Lattice"]
         super(CubBayesLatticeG,self).__init__(distribution, integrand, allowed_levels, allowed_distribs)
 
@@ -141,38 +146,27 @@ class CubBayesLatticeG(StoppingCriterion):
             raise ParameterError("CubLattice_g requires distribution to have randomize=False")
         if distribution.backend != 'gail':
             raise ParameterError("CubBayesLattice_g requires distribution to have 'GAIL' backend")
-        # Construct AccumulateData Object to House Integration data
-        # self.data = LDTransformBayesData(self, integrand, self.fft_update, m_min, m_max, fudge, check_cone)
-        self.distribution = distribution
 
     # computes the integral
     def integrate(self):
-        tstart = time()  # start the timer
-        n = 2 ** self.mvec[0]
-        order_ = None
-        stop_flag = None
-        muhat = None
 
         # pick a random value to apply as shift
         shift = np.random.rand(1, self.dim)
-
         # shift = np.array([[0.58, 0.632]])
 
-        xun_ = np.array([])
-        xpts_ = np.array([])
-        ftilde_ = np.array([])  # temporary storage between iterations
+        # Construct AccumulateData Object to House Integration data
+        self.data = LDTransformBayesData(self, self.integrand, self.m_min, self.m_max, shift, self.vdc_order)
+
+        tstart = time()  # start the timer
+
         # Iteratively find the number of points required for the cubature to meet
         # the error threshold
-        for iter, m in enumerate(self.mvec):
+        while True:
             tstart_iter = time()
-            n = 2 ** m
 
             # Update function values
-            if self.vdc_order:
-                ftilde_, xun_, xpts_ = self.iter_fft_vdc(iter, shift, xun_, xpts_, ftilde_)
-            else:
-                ftilde_, xun_, xpts_ = self.iter_fft(iter, shift, xun_, xpts_, ftilde_)
-            stop_flag, muhat, order_ = self.stopping_criterion(xun_, ftilde_, iter, m)
+            xun_, ftilde_, m = self.data.update_data()
+            stop_flag, muhat, order_ = self.stopping_criterion(xun_, ftilde_, m)
 
             self.timeAll.append(time() - tstart_iter)  # store per iteration time
 
@@ -182,15 +176,21 @@ class CubBayesLatticeG(StoppingCriterion):
             if self.stopAtTol and stop_flag:
                 break
 
+            if m >= self.m_max:
+                warnings.warn(f'Already used maximum allowed sample size {2 ** self.m_max}.'
+                              f' Note that error tolerances may no longer be satisfied',
+                              MaxSamplesWarning)
+                break
+
         out = OutParams()
-        out.n = n
+        out.n = 2**m
         out.time = tstart - time()
         out.ErrBd = self.errorBdAll[-1]
 
         optParams = OutOptParams()
         optParams.ErrBdAll = self.errorBdAll
         optParams.muhatAll = self.muhatAll
-        optParams.mvec = self.mvec
+        optParams.mvec = self.data.mvec
         optParams.aMLEAll = self.aMLEAll
         optParams.timeAll = self.timeAll
         optParams.s_All = self.s_All
@@ -205,96 +205,11 @@ class CubBayesLatticeG(StoppingCriterion):
             out.exitflag = 1
         else:
             out.exitflag = 2  # error tolerance may not be met
-        if stop_flag == False:
-            warnings.warn(f'In order to achieve the guaranteed accuracy, used maximum allowed sample size {n}',
-                          MaxSamplesWarning)
+
         return muhat, out
 
-    # Efficient FFT computation algorithm, avoids recomputing the full fft
-    def iter_fft(self, iter, shift, xun, xpts, ftildePrev):
-        m = self.mvec[iter]
-        n = 2 ** m
-
-        # In every iteration except the first one, "n" number_of_points is doubled,
-        # but FFT is only computed for the newly added points.
-        # Previously computed FFT is reused.
-        if iter == 0:
-            # In the first iteration compute full FFT
-            # xun_ = mod(bsxfun( @ times, (0:1 / n:1-1 / n)',self.gen_vec),1)
-            xun_ = np.arange(0, 1, 1 / n).reshape((n, 1))
-            xun_ = np.mod((xun_ * self.gen_vec), 1)
-
-            xun_ = self.distribution.gen_samples(n_min=0, n_max=n)
-
-            # xpts_ = np.mod(bsxfun( @ plus, xun_, shift), 1)  # shifted
-            xpts_ = np.mod((xun_ + shift), 1)  # shifted
-
-            # Compute initial FFT
-            ftilde_ = np.fft.fft(self.ff(xpts_))  # evaluate integrand's fft
-            ftilde_ = ftilde_.reshape((n, 1))
-        else:
-            # xunnew = np.mod(bsxfun( @ times, (1/n : 2/n : 1-1/n)',self.gen_vec),1)
-            xunnew = np.arange(1 / n, 1, 2 / n).reshape((n//2, 1))
-            xunnew = np.mod(xunnew * self.gen_vec, 1)
-
-            xunnew = self.distribution.gen_samples(n_min=n//2, n_max=n)
-
-            # xnew = np.mod(bsxfun( @ plus, xunnew, shift), 1)
-            xnew = np.mod((xunnew + shift), 1)
-
-            [xun_, xpts_] = self.merge_pts(xun, xunnew, xpts, xnew, n, self.dim)
-            mnext = m - 1
-
-            # Compute FFT on next set of new points
-            ftildeNextNew = np.fft.fft(self.ff(xnew))
-            ftildeNextNew = ftildeNextNew.reshape((n//2, 1))
-            if self.debugEnable:
-                CubBayesLatticeG.alertMsg(ftildeNextNew, 'Nan', 'Inf')
-
-            # combine the previous batch and new batch to get FFT on all points
-            ftilde_ = self.merge_fft(ftildePrev, ftildeNextNew, mnext)
-
-        return ftilde_, xun_, xpts_
-
-    # Lattice points are ordered in van der Corput sequence, so we cannot use
-    # Matlab's built-in fft routine. We use a custom fft instead.
-    def iter_fft_vdc(self, iter, shift, xun, xpts, ftildePrev):
-        m = self.mvec[iter]
-        n = 2 ** m
-
-        # In every iteration except the first one, "n" number_of_points is doubled,
-        # but FFT is only computed for the newly added points.
-        # Previously computed FFT is reused.
-        if iter == 1:
-            # in the first iteration compute the full FFT
-            [xpts_, xun_] = self.simple_lattice_gen(n, self.dim, shift, True)
-
-            # Compute initial FFT
-            ftilde_ = self.fft_DIT(self.ff(xpts_), m)  # evaluate integrand's fft
-        else:
-            [xnew, xunnew] = self.simple_lattice_gen(n, self.dim, shift, False)
-            mnext = m - 1
-
-            # Compute FFT on next set of new points
-            ftildeNextNew = self.fft_DIT(self.ff(xnew), mnext)
-            if self.debugEnable:
-                CubBayesLatticeG.alertMsg(ftildeNextNew, 'Nan', 'Inf')
-
-            xpts_ = np.vstack(xpts, xnew)
-            temp = np.zeros(n, self.dim)
-            temp[0:: 2, :] = xun
-            temp[1:: 2, :] = xunnew
-            xun_ = temp
-            # combine the previous batch and new batch to get FFT on all points
-            ftilde_ = self.merge_fft(ftildePrev, ftildeNextNew, mnext)
-
-        if self.debugEnable:
-            CubBayesLatticeG.alertMsg(ftilde_, 'Inf', 'Nan')
-
-        return ftilde_, xun_, xpts_
-
     # decides if the user-defined error threshold is met
-    def stopping_criterion(self, xpts, ftilde, iter, m):
+    def stopping_criterion(self, xpts, ftilde, m):
         ftilde = ftilde.squeeze()
         n = 2 ** m
         success = False
@@ -408,76 +323,13 @@ class CubBayesLatticeG(StoppingCriterion):
             loss = loss1 + loss2
 
         if self.debugEnable:
-            self.alertMsg(loss1, 'Inf', 'Imag')
-            self.alertMsg(RKHSnorm, 'Imag')
-            self.alertMsg(loss2, 'Inf', 'Imag')
-            self.alertMsg(loss, 'Inf', 'Imag', 'Nan')
-            self.alertMsg(Lambda, 'Imag')
+            self.data.alertMsg(loss1, 'Inf', 'Imag')
+            self.data.alertMsg(RKHSnorm, 'Imag')
+            self.data.alertMsg(loss2, 'Inf', 'Imag')
+            self.data.alertMsg(loss, 'Inf', 'Imag', 'Nan')
+            self.data.alertMsg(Lambda, 'Imag')
 
         return loss, Lambda, Lambda_ring, RKHSnorm
-
-    # prints debug message if the given variable is Inf, Nan or complex, etc
-    # Example: alertMsg(x, 'Inf', 'Imag')
-    #          prints if variable 'x' is either Infinite or Imaginary
-    @staticmethod
-    def alertMsg(*args):
-        varargin = args
-        nargin = len(varargin)
-        if nargin > 1:
-            iStart = 0
-            varTocheck = varargin[iStart]
-            iStart = iStart + 1
-            inpvarname = 'variable'
-
-            while iStart < nargin:
-                type = varargin[iStart]
-                iStart = iStart + 1
-
-                if type == 'Nan':
-                    if np.any(np.isnan(varTocheck)):
-                        print(f'{inpvarname} has NaN values')
-                elif type == 'Inf':
-                    if np.any(np.isinf(varTocheck)):
-                        print(f'{inpvarname} has Inf values')
-                elif type == 'Imag':
-                    if not np.all(np.isreal(varTocheck)):
-                        print(f'{inpvarname} has complex values')
-                else:
-                    print('unknown type check requested !')
-
-    # computes the periodization transform for the given function values
-    @staticmethod
-    def doPeriodTx(fInput, ptransform):
-
-        if ptransform == 'Baker':
-            f = lambda x: fInput(1 - 2 * abs(x - 1 / 2))  # Baker's transform
-        elif ptransform == 'C0':
-            f = lambda x: fInput(3 * x ** 2 - 2 * x ** 3) * np.prod(6 * x * (1 - x), 1)  # C^0 transform
-        elif ptransform == 'C1':
-            # C^1 transform
-            f = lambda x: fInput(x ** 3 * (10 - 15 * x + 6 * x ** 2)) * np.prod(30 * x ** 2 * (1 - x) ** 2, 1)
-        elif ptransform == 'C1sin':
-            # Sidi C^1 transform
-            f = lambda x: fInput(x - np.sin(2 * np.pi * x) / (2 * np.pi)) * np.prod(2 * np.sin(np.pi * x) ** 2, 1)
-        elif ptransform == 'C2sin':
-            # Sidi C^2 transform
-            psi3 = lambda t: (8 - 9 * np.cos(np.pi * t) + np.cos(3 * np.pi * t)) / 16
-            psi3_1 = lambda t: (9 * np.sin(np.pi * t) * np.pi - np.sin(3 * np.pi * t) * 3 * np.pi) / 16
-            f = lambda x: fInput(psi3(x)) * np.prod(psi3_1(x), 1)
-        elif ptransform == 'C3sin':
-            # Sidi C^3 transform
-            psi4 = lambda t: (12 * np.pi * t - 8 * np.sin(2 * np.pi * t) + np.sin(4 * np.pi * t)) / (12 * np.pi)
-            psi4_1 = lambda t: (12 * np.pi - 8 * np.cos(2 * np.pi * t) * 2 * np.pi + np.sin(
-                4 * np.pi * t) * 4 * np.pi) / (12 * np.pi)
-            f = lambda x: fInput(psi4(x)) * np.prod(psi4_1(x), 1)
-        elif ptransform == 'none':
-            # do nothing
-            f = lambda x: fInput(x)
-        else:
-            f = fInput
-            print(f'Error: Periodization transform {ptransform} not implemented')
-
-        return f
 
     # Computes modified kernel Km1 = K - 1
     # Useful to avoid cancellation error in the computation of (1 - n/\lambda_1)
@@ -553,104 +405,3 @@ class CubBayesLatticeG(StoppingCriterion):
             Lambda_ring = 0
 
         return Lambda, Lambda_ring
-
-    # just returns the generator for rank-1 Lattice point generation
-    @staticmethod
-    def get_lattice_gen_vec(d):
-        z = [1, 433461, 315689, 441789, 501101, 146355, 88411, 215837, 273599,
-             151719, 258185, 357967, 96407, 203741, 211709, 135719, 100779,
-             85729, 14597, 94813, 422013, 484367]  # generator
-        z = np.array(z[:d]).reshape((1, d))
-        return z
-
-    # generates rank-1 Lattice points in Vander Corput sequence order
-    @staticmethod
-    def simple_lattice_gen(n, d, shift, firstBatch):
-        z = CubBayesLatticeG.get_lattice_gen_vec(d)
-
-        nmax = n
-        nmin = 1 + n / 2
-        if firstBatch == True:
-            nmin = 1
-
-        nelem = nmax - nmin + 1
-
-        if firstBatch == True:
-            brIndices = CubBayesLatticeG.vdc(nelem).transpose()
-
-            # xlat_ = np.mod(bsxfun( @ times, (0:1 / n:1-1 / n)',z),1) # unshifted in direct order
-            xlat_ = np.mod((np.arange(0, 1 - 1 / n, 1 / n).transpose() * z), 1)
-        else:
-            brIndices = CubBayesLatticeG.vdc(nelem).transpose() + 1 / (2 * (nmin - 1))
-            # xlat_ = mod(bsxfun( @ times, (1 / n:2 / n:1-1 / n)',z),1) # unshifted in direct order
-            xlat_ = np.mod((np.arange(1 / n, 1 - 1 / n, 2 / n).transpose() * z), 1)
-
-        # xlat = mod(bsxfun( @ times, brIndices',z),1)  # unshifted
-        xlat = np.mod((brIndices.transpose() * z), 1)
-        # xlat = mod(bsxfun( @ plus, xlat, shift), 1)  # shifted in VDC order
-        xlat = np.mod((xlat + shift), 1)
-
-        return [xlat, xlat_]
-
-    # van der Corput sequence in base 2
-    @staticmethod
-    def vdc(n):
-        if n > 1:
-            k = np.log2(n)  # We compute the VDC seq part by part of power 2 size
-            q = np.zeros(2 ** k, 1)
-            for l in range(0, k - 1):
-                nl = 2 ** l
-                kk = 2 ** (k - l - 1)
-                ptind = np.matlib.repmat(np.vstack(np.full((nl, 1), False), np.full((nl, 1), True)), kk, 1)
-                q[ptind] = q[ptind] + 1 / 2 ** (l + 1)
-        else:
-            q = 0
-        return q
-
-    # fft with decimation in time i.e., input is already in 'bitrevorder'
-    @staticmethod
-    def fft_DIT(y, nmmin):
-        for l in range(0, nmmin - 1):
-            nl = 2 ** l
-            nmminlm1 = 2 ** (nmmin - l - 1)
-            # ptind=repmat([true(nl,1); false(nl,1)], nmminlm1,1)
-            ptind = np.matlib.repmat(np.vstack(np.full((nl, 1), True), np.full((nl, 1), False)), nmminlm1, 1)
-            # coef=exp(-2*pi()*sqrt(-1)*(0:nl-1)'/(2*nl))
-            coef = exp(-2 * np.pi * 1j * np.arange(0, nl - 1).transpose() / (2 * nl))
-            coefv = np.matlib.repmat(coef, nmminlm1, 1)
-            evenval = y[ptind]
-            oddval = y[~ptind]
-            y[ptind] = (evenval + coefv * oddval)
-            y[~ptind] = (evenval - coefv * oddval)
-        return y
-
-    # using FFT butefly plot technique merges two halves of fft
-    @staticmethod
-    def merge_fft(ftildeNew, ftildeNextNew, mnext):
-        ftildeNew = np.vstack([ftildeNew, ftildeNextNew])
-        nl = 2 ** mnext
-        # ptind=[true(nl,1); false(nl,1)]
-        ptind = np.ndarray(shape=(2 * nl, 1), buffer=np.array([True] * nl + [False] * nl), dtype=bool)
-        # coef = exp(-2*1j*(0:nl-1)'/(2*nl))
-        coef = exp(-2 * np.pi * 1j * np.ndarray(shape=(nl, 1), buffer=np.arange(0, nl), dtype=int) / (2 * nl))
-        # coefv = np.matlib.repmat(coef, 1, 1)
-        coefv = np.tile(coef, (1, 1))
-        evenval = ftildeNew[ptind].reshape((nl, 1))
-        oddval = ftildeNew[~ptind].reshape((nl, 1))
-        ftildeNew[ptind] = np.squeeze(evenval + coefv * oddval)
-        ftildeNew[~ptind] = np.squeeze(evenval - coefv * oddval)
-        return ftildeNew
-
-    # inserts newly generated points with the old set by interleaving them
-    # xun - unshifted points
-    @staticmethod
-    def merge_pts(xun, xunnew, x, xnew, n, d):
-        temp = np.zeros((n, d))
-        temp[0::2, :] = xun
-        temp[1::2, :] = xunnew
-        xun = temp
-        temp = np.zeros((n, d))
-        temp[0::2, :] = x
-        temp[1::2, :] = xnew
-        x = temp
-        return xun, x
